moved contents of MapScreen to Explore
