--- app/src/main/java/com/github/se/gomeet/ui/MapScreen.kt
@@ -1,50 +1,0 @@
-package com.github.se.gomeet.ui
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.tooling.preview.Preview
-import com.google.maps.android.compose.CameraPositionState
-import com.google.maps.android.compose.GoogleMap
-import com.google.maps.android.compose.MapProperties
-import com.google.maps.android.compose.MapType
-import com.google.maps.android.compose.MapUiSettings
-import com.google.maps.android.compose.rememberCameraPositionState
-
-@Preview
-@Composable
-fun MapScreen() {
-  Box(Modifier.fillMaxSize().testTag("MapScreen")) { GoogleMapView() }
-}
-
-@Composable
-fun GoogleMapView(
-    modifier: Modifier = Modifier,
-    cameraPositionState: CameraPositionState = rememberCameraPositionState(),
-    onMapLoaded: () -> Unit = {},
-    content: @Composable () -> Unit = {},
-) {
-
-  val uiSettings by remember {
-    mutableStateOf(MapUiSettings(compassEnabled = false, zoomControlsEnabled = false))
-  }
-  val mapProperties by remember { mutableStateOf(MapProperties(mapType = MapType.NORMAL)) }
-  val mapVisible by remember { mutableStateOf(true) }
-
-  if (mapVisible) {
-    GoogleMap(
-        modifier = modifier,
-        cameraPositionState = cameraPositionState,
-        properties = mapProperties,
-        uiSettings = uiSettings,
-        onMapLoaded = onMapLoaded,
-        onPOIClick = {}) {
-          content()
-        }
-  }
-}
