+++ app/src/main/java/com/github/se/gomeet/ui/explore/Explore.kt
@@ -1,28 +1,65 @@
 package com.github.se.gomeet.ui.explore
 
 import android.util.Log
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import com.github.se.gomeet.ui.navigation.BottomNavigationMenu
 import com.github.se.gomeet.ui.navigation.NavigationActions
 import com.github.se.gomeet.ui.navigation.Route
 import com.github.se.gomeet.ui.navigation.TOP_LEVEL_DESTINATIONS
+import com.google.maps.android.compose.CameraPositionState
+import com.google.maps.android.compose.GoogleMap
+import com.google.maps.android.compose.MapProperties
+import com.google.maps.android.compose.MapType
+import com.google.maps.android.compose.MapUiSettings
+import com.google.maps.android.compose.rememberCameraPositionState
 
 @Composable
 fun Explore(nav: NavigationActions) {
-  Log.d("Explore", "Back in Explore")
-  Scaffold(
-      bottomBar = {
-        BottomNavigationMenu(
-            onTabSelect = { selectedTab ->
-              nav.navigateTo(TOP_LEVEL_DESTINATIONS.first { it.route == selectedTab })
-            },
-            tabList = TOP_LEVEL_DESTINATIONS,
-            selectedItem = Route.EXPLORE)
-      }) { innerPadding ->
-        Text("Explore", Modifier.padding(innerPadding))
-      }
+    Log.d("Explore", "Back in Explore")
+    Scaffold(
+        bottomBar = {
+            BottomNavigationMenu(
+                onTabSelect = { selectedTab ->
+                    nav.navigateTo(TOP_LEVEL_DESTINATIONS.first { it.route == selectedTab })
+                },
+                tabList = TOP_LEVEL_DESTINATIONS,
+                selectedItem = Route.EXPLORE)
+        }
+    ) {
+        innerPadding -> Box(modifier = Modifier.padding(innerPadding)) { GoogleMapView() }
+    }
+}
+
+@Composable
+fun GoogleMapView(
+    modifier: Modifier = Modifier,
+    cameraPositionState: CameraPositionState = rememberCameraPositionState(),
+    onMapLoaded: () -> Unit = {},
+    content: @Composable () -> Unit = {},
+) {
+
+    val uiSettings by remember {
+        mutableStateOf(MapUiSettings(compassEnabled = false, zoomControlsEnabled = false))
+    }
+    val mapProperties by remember { mutableStateOf(MapProperties(mapType = MapType.NORMAL)) }
+    val mapVisible by remember { mutableStateOf(true) }
+
+    if (mapVisible) {
+        GoogleMap(
+            modifier = modifier,
+            cameraPositionState = cameraPositionState,
+            properties = mapProperties,
+            uiSettings = uiSettings,
+            onMapLoaded = onMapLoaded,
+            onPOIClick = {}) {
+            content()
+        }
+    }
 }
